+++ server/app/src/router/api/minihome/minihome.repository.ts
@@ -1,7 +1,7 @@
 import { Kysely } from "kysely";
 
 import { DataBase } from "../../../types/database";
-import { addTimeStamp } from "../../../utils/addTimeStamp";
+import { addTimeStamp } from "../../../utils/timestamp";
 import { MinihomeTable } from "./minihome.schema";
 
 export class MinihomeRepository {
